ecommerce/requirements.txt
aioredis==1.3.1
arrow==0.15.4
asgi-redis==1.4.3
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
certifi==2019.11.28
cffi==1.13.2
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2
django-allauth==0.41.0
django-crispy-forms==1.8.1
django-filter==2.2.0
django-heroku==0.3.1
django-storages==1.8
djangorestframework==3.11.0
dropbox==9.4.0
gunicorn==20.0.4
h2==3.1.1
hiredis==1.0.1
hpack==3.0.0
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
Jinja2==2.10.3
jinja2-time==0.2.0
make==0.1.5
Markdown==3.1.1
MarkupSafe==1.1.1
msgpack==0.6.2
msgpack-python==0.5.6
oauthlib==3.1.0
Pillow==6.2.1
priority==1.3.0
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyHamcrest==1.9.0
PyMySQL==0.9.3
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
redis==2.10.6
requests==2.22.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.13.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==18.8.1
urllib3==1.25.7
whitenoise==5.0.1
zope.interface==4.7.1
